Restructure system tests package
