+++ cloud.foundry.cli/src/test/java/cloud/foundry/cli/system/StandardErrCapturingTest.java
@@ -1,5 +1,7 @@
 package cloud.foundry.cli.system;
 
+import cloud.foundry.cli.system.util.ArgumentsBuilder;
+import cloud.foundry.cli.system.util.RunResult;
 import org.junit.jupiter.api.Test;
