+++ cloud.foundry.cli/src/test/java/cloud/foundry/cli/system/util/ArgumentsBuilder.java
@@ -1,4 +1,4 @@
-package cloud.foundry.cli.system;
+package cloud.foundry.cli.system.util;
 
 import java.util.ArrayList;
 import java.util.List;
