+++ cloud.foundry.cli/src/test/java/cloud/foundry/cli/system/util/StreamContents.java
@@ -1,4 +1,4 @@
-package cloud.foundry.cli.system;
+package cloud.foundry.cli.system.util;
 
 
 /**
